pyproject.toml
[project]
name = "kindle-clippings-parser"
version = "0.1.0"
description = "A parser for Amazon Kindle clippings files that converts highlights and notes into organized markdown format"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dateutil>=2.8.0",
]

[project.scripts]
kindle-parser = "parser:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["parser", "utils"]
